+++ src/harvestine/types.h
@@ -4,22 +4,22 @@
 // Types
 // --------------------------------------
 
-typedef signed char     i8;
-typedef unsigned char   u8;
-typedef short           i16;
-typedef unsigned short  u16;
-typedef int             i32;
-typedef unsigned int    u32;
-typedef long            i64;
-typedef unsigned long   u64;
-typedef float           f32;
-typedef double          f64;
+typedef signed char         i8;
+typedef unsigned char       u8;
+typedef short               i16;
+typedef unsigned short      u16;
+typedef int                 i32;
+typedef unsigned int        u32;
+typedef long                i64;
+typedef unsigned long long  u64;
+typedef float               f32;
+typedef double              f64;
 
-typedef i32             b32;
-#define true            1
-#define false           0
+typedef i32                 b32;
+#define true                1
+#define false               0
 
-#define U64_MAX         -1UL
+#define U64_MAX             -1UL
 
 // --------------------------------------
 // Compiler
